+++ src/main/java/blusunrize/immersiveengineering/client/ClientEventHandler.java
@@ -46,7 +46,7 @@ import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;
 import net.minecraft.ChatFormatting;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.Font.DisplayMode;
+import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.model.EntityModel;
 import net.minecraft.client.model.HeadedModel;
 import net.minecraft.client.multiplayer.ClientLevel;
@@ -99,6 +99,7 @@ import java.util.Optional;
 import java.util.stream.IntStream;
 
 import static blusunrize.immersiveengineering.ImmersiveEngineering.rl;
+import static blusunrize.immersiveengineering.api.IEApi.ieLoc;
 
 public class ClientEventHandler implements ResourceManagerReloadListener
 {
@@ -306,78 +307,76 @@ public class ClientEventHandler implements ResourceManagerReloadListener
 	@SubscribeEvent
 	public void onRenderOverlayPre(RenderGuiLayerEvent.Pre event)
 	{
-		if(!ZoomHandler.isZooming||!event.getLayer().equals(VanillaGuiLayers.CROSSHAIR))
+		ItemStack equipped = ClientUtils.mc().player.getItemInHand(InteractionHand.MAIN_HAND);
+		// early exit if not handling a zoom tool
+		if(!(event.getName().equals(VanillaGuiLayers.CROSSHAIR)&&ZoomHandler.isZooming&&equipped.getItem() instanceof IZoomTool tool))
 			return;
-
 		event.setCanceled(true);
-		MultiBufferSource.BufferSource buffers = event.getGuiGraphics().bufferSource();
-		PoseStack transform = new PoseStack();
+
+		GuiGraphics graphics = event.getGuiGraphics();
+		PoseStack transform = graphics.pose();
 		transform.pushPose();
 		int width = ClientUtils.mc().getWindow().getGuiScaledWidth();
 		int height = ClientUtils.mc().getWindow().getGuiScaledHeight();
 		int resMin = Math.min(width, height);
-		float offsetX = (width-resMin)/2f;
-		float offsetY = (height-resMin)/2f;
+		float offsetX = Math.round((width-resMin)/2f);
+		float offsetY = Math.round((height-resMin)/2f);
 
 		if(resMin==width)
 		{
-			GuiHelper.drawColouredRect(0, 0, width, (int)offsetY+1, 0xff000000, buffers, transform);
-			GuiHelper.drawColouredRect(0, (int)offsetY+resMin, width, (int)offsetY+1, 0xff000000, buffers, transform);
+			graphics.fill(0, 0, width, (int)offsetY+1, 0xff000000);
+			graphics.fill(0, (int)offsetY+resMin, width, (int)(offsetY+resMin+offsetY+1), 0xff000000);
 		}
 		else
 		{
-			GuiHelper.drawColouredRect(0, 0, (int)offsetX+1, height, 0xff000000, buffers, transform);
-			GuiHelper.drawColouredRect((int)offsetX+resMin, 0, (int)offsetX+1, height, 0xff000000, buffers, transform);
+			graphics.fill(0, 0, (int)offsetX+1, height, 0xff000000);
+			graphics.fill((int)offsetX+resMin, 0, (int)(offsetX+resMin+offsetX+1), height, 0xff000000);
 		}
 		transform.translate(offsetX, offsetY, 0);
-		VertexConsumer builder = buffers.getBuffer(IERenderTypes.getGuiTranslucent(rl("textures/gui/scope.png")));
-		GuiHelper.drawTexturedColoredRect(builder, transform, 0, 0, resMin, resMin, 1, 1, 1, 1, 0f, 1f, 0f, 1f);
 
-		builder = buffers.getBuffer(IERenderTypes.getGui(rl("textures/gui/hud_elements.png")));
-		GuiHelper.drawTexturedColoredRect(builder, transform, 218/256f*resMin, 64/256f*resMin, 24/256f*resMin, 128/256f*resMin, 1, 1, 1, 1, 64/256f, 88/256f, 96/256f, 224/256f);
-		ItemStack equipped = ClientUtils.mc().player.getItemInHand(InteractionHand.MAIN_HAND);
-		if(!equipped.isEmpty()&&equipped.getItem() instanceof IZoomTool tool)
+		RenderSystem.enableBlend();
+		graphics.blitSprite(ieLoc("hud/scope"), 0, 0, resMin, resMin);
+		RenderSystem.disableBlend();
+
+		float[] steps = tool.getZoomSteps(equipped, ClientUtils.mc().player);
+		if(steps!=null&&steps.length > 1)
 		{
-			float[] steps = tool.getZoomSteps(equipped, ClientUtils.mc().player);
-			if(steps!=null&&steps.length > 1)
+			// draw gauge on the right side
+			transform.translate(218/256f*resMin, 64/256f*resMin, 0);
+			graphics.blitSprite(ieLoc("hud/gauge_vertical"), 0, 0, Math.round(24/256f*resMin), Math.round(128/256f*resMin));
+
+			float stepLength = 118/(float)steps.length;
+			float stepOffset = (stepLength-7)/2f;
+			// move inside the gauge
+			RenderSystem.enableBlend();
+			transform.translate(5/256f*resMin, (5+stepOffset)/256f*resMin, 0);
+
+			// draw markers for the possible steps
+			int curStep = -1;
+			float dist = 0;
+			int innerWidth = Math.round(14/256f*resMin);
+			int innerHeight = Math.round(7/256f*resMin);
+
+			for(int i = 0; i < steps.length; i++)
 			{
-				int curStep = -1;
-				float dist = 0;
-
-				float totalOffset = 0;
-				float stepLength = 118/(float)steps.length;
-				float stepOffset = (stepLength-7)/2f;
-				transform.translate(223/256f*resMin, 64/256f*resMin, 0);
-				transform.translate(0, (5+stepOffset)/256*resMin, 0);
-				for(int i = 0; i < steps.length; i++)
+				graphics.blitSprite(ieLoc("hud/gauge_vertical_step"), 0, 0, innerWidth, innerHeight);
+				transform.translate(0, stepLength/256*resMin, 0);
+				if(curStep==-1||Math.abs(steps[i]-ZoomHandler.fovZoom) < dist)
 				{
-					GuiHelper.drawTexturedColoredRect(builder, transform, 0, 0, 8/256f*resMin, 7/256f*resMin, 1, 1, 1, 1, 88/256f, 96/256f, 96/256f, 103/256f);
-					transform.translate(0, stepLength/256*resMin, 0);
-					totalOffset += stepLength;
-
-					if(curStep==-1||Math.abs(steps[i]-ZoomHandler.fovZoom) < dist)
-					{
-						curStep = i;
-						dist = Math.abs(steps[i]-ZoomHandler.fovZoom);
-					}
+					curStep = i;
+					dist = Math.abs(steps[i]-ZoomHandler.fovZoom);
 				}
-				transform.translate(0, -totalOffset/256*resMin, 0);
-
-				if(curStep < steps.length)
-				{
-					transform.translate(6/256f*resMin, curStep*stepLength/256*resMin, 0);
-					GuiHelper.drawTexturedColoredRect(builder, transform, 0, 0, 8/256f*resMin, 7/256f*resMin, 1, 1, 1, 1, 88/256f, 98/256f, 103/256f, 110/256f);
-					ClientUtils.font().drawInBatch((1/steps[curStep])+"x", (int)(16/256f*resMin), 0, 0xffffff, true,
-							transform.last().pose(), buffers, DisplayMode.NORMAL, 0, 0xf000f0);
-					transform.translate(-6/256f*resMin, -curStep*stepLength/256*resMin, 0);
-				}
-				transform.translate(0, -((5+stepOffset)/256*resMin), 0);
-				transform.translate(-223/256f*resMin, -64/256f*resMin, 0);
 			}
-		}
+			transform.translate(0, -118/256f*resMin, 0);
 
-		transform.translate(-offsetX, -offsetY, 0);
-		buffers.endBatch();
+			// draw the pointer and text for current level
+			transform.translate(0, curStep*stepLength/256*resMin, 0);
+			graphics.blitSprite(ieLoc("hud/gauge_vertical_pointer"), 0, 0, innerWidth, innerHeight);
+			transform.translate(16/256f*resMin, 1/256f*resMin, 0);
+			graphics.drawString(ClientUtils.font(), (1/steps[curStep])+"x", 0, 0, 0xffffff, false);
+			RenderSystem.disableBlend();
+		}
+		transform.popPose();
 	}
 
 	@SubscribeEvent()